+++ CHANGES.rst
@@ -4,10 +4,12 @@
 New Features
 ------------
 
-- Adding flux/surface brightness translation and surface brightness
+- Added flux/surface brightness translation and surface brightness
   unit conversion in Cubeviz and Specviz. [#2781]
 
 - Plugin tray is now open by default. [#2892]
+
+- New "About" plugin to show Jdaviz version info. [#2886]
 
 Cubeviz
 ^^^^^^^